+++ src/db/db.rs
@@ -81,9 +81,12 @@ pub trait DbConnection {
     fn delete_old_clicks(&mut self, seconds: u64) -> Result<(), Box<dyn Error>>;
     fn delete_removed_from_history(&mut self) -> Result<(), Box<dyn Error>>;
     fn delete_unused_streaming_servers(&mut self, seconds: u64) -> Result<(), Box<dyn Error>>;
+    fn detect_language_codes(&mut self, language_to_code: &HashMap<String,String>) -> Result<u32, Box<dyn Error>>;
     fn remove_unused_ip_infos_from_stationclicks(&mut self, seconds: u64) -> Result<(), Box<dyn Error>>;
     fn remove_illegal_icon_links(&mut self) -> Result<(), Box<dyn Error>>;
     fn calc_country_field(&mut self) -> Result<(), Box<dyn Error>>;
+
+    fn replace_languages(&mut self, items: &HashMap<String,String>) -> Result<u32, Box<dyn Error>>;
     
     fn update_stations_clickcount(&self) -> Result<(), Box<dyn Error>>;
     fn clean_urls(&self, table_name: &str, column_key: &str, column_url: &str, allow_empty: bool) -> Result<(), Box<dyn Error>>;